refactor: (dpk) clean the code and add edge cases
I have assumed that, since there are many conditionals,
the data should be from a non-trusted source. The ideal
would be having Typescript, because making so many checks
during runtime execution increases the complexity. Specially
because Javascript does not has a consistent way to verify
types - `typeof [1, 2, 3]` or `typeof new Date()` or
`typeof { clipboard: 'health' }` will end up with the same
result `object`. Also, I have interpred "clean code" and
"readable" as a balance between "easy to understand what is
happening" and "short and efficcient".

In this case scenario, I choose to make short circuits, verifying
the most shallow and known cases (determined by the constants)
before the most deep and uncertain cases - the checks are required
to avoid bugs and unexpected behaviors (like some edge cases not
considered in the existing code), and this approach helps to reduce
time and complexity in the first cases. I have splitted the code in
minor and testable functions to avoid code repetition and to absctract
the inner workings, letting the clarity to the function name - just
by reading inside the function `deterministicPartitionKey` is possible
to know the rules.